requirements.txt
fastapi
uvicorn
openai-whisper
language-tool-python
librosa
noisereduce
pydub
soundfile
google-generativeai
boto3
pyjwt
jose
streamlit
torch==2.3.0
streamlit
language-tool-python
librosa
noisereduce
soundfile
google-generativeai
pydub
openai-whisper
